+++ frontend/src/main/java/org/example/frontend/data/models/TourLogInput.java
@@ -6,7 +6,7 @@ public record TourLogInput(
         Long timestamp,
         String comment,
         Integer difficulty,
-        Long distance,
-        Integer duration,
+        Integer distance,
+        Long duration,
         Integer rating) {
 }
